트렌드/mcp_online/mcp_server/requirements.txt
fastmcp==2.11.2
openai==1.42.0
pdfkit==1.0.0
python-dotenv==1.1.0
jinja2==3.1.2
PyMySQL==1.1.0